+++ src/lib/errors/graphqlErrors.ts
@@ -1,9 +1,13 @@
 import {GraphQLError} from "graphql";
-import {BaseError} from "./BaseError";
 
 export class NotLoggedInGqlError extends GraphQLError {
-
     constructor() {
         super("Not logged in");
+    }
+}
+
+export class PostNotFoundGqlError extends GraphQLError {
+    constructor(postId: number) {
+        super(`Post '${postId}' not found!`);
     }
 }