+++ src/lib/globals.ts
@@ -9,7 +9,6 @@ import {EventEmitter} from "events";
 import * as fsx from "fs-extra";
 import * as yaml from "js-yaml";
 import * as winston from "winston";
-import {MemoryCache} from "./MemoryCache";
 
 const configPath = "config.yaml";
 const defaultConfig = __dirname + "/../default-config.yaml";
@@ -28,14 +27,13 @@ if (!(fsx.pathExistsSync(configPath))) {
  */
 namespace globals {
     export const config = yaml.safeLoad(fsx.readFileSync("config.yaml", "utf-8"));
-    export const cache = new MemoryCache(1200);
     export const logger = winston.createLogger({
         transports: [
             new winston.transports.Console({
                 format: winston.format.combine(
                     winston.format.timestamp(),
                     winston.format.colorize(),
-                    winston.format.printf(({ level, message, timestamp }) => {
+                    winston.format.printf(({level, message, timestamp}) => {
                         return `${timestamp} ${level}: ${message}`;
                     }),
                 ),
@@ -44,9 +42,6 @@ namespace globals {
         ],
     });
     export const internalEmitter = new EventEmitter();
-    cache.on("set", (key) => logger.debug(`Caching '${key}'.`));
-    cache.on("miss", (key) => logger.debug(`Cache miss for '${key}'`));
-    cache.on("hit", (key) => logger.debug(`Cache hit for '${key}'`));
 }
 
 export default globals;