+++ CHANGELOG.md
@@ -8,10 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Connection to Postgres Database
 - Graphql Schema
 - default-config file and generation of config file on startup
 - DTOs
 - Home Route
-- database caching
 - session management
+- Sequelize models and integration
+- Sequelize-typescript integration